Remove owned data flag for now
The flag adds a fair bit of complexity, and we're not sure we'll even
need it. Indeed we're going to have to copy from C ordered data in
Python to Fortran ordered anyway, so we're going to have to copy on the
way in and the way out.

This backs it out for now, we can always add it back in if we think we
need it.